Add TCL 2020 and carbon layers to treecoverloss analysis, remove annualupdate carbonflux_minimal analyses
