--- src/main/scala/org/globalforestwatch/summarystats/annualupdate/AnnualUpdateAnalysis.scala
@@ -1,39 +1,0 @@
-package org.globalforestwatch.summarystats.annualupdate
-
-import java.time.LocalDateTime
-import java.time.format.DateTimeFormatter
-
-import geotrellis.vector.{Feature, Geometry}
-import org.apache.spark.rdd.RDD
-import org.apache.spark.sql.SparkSession
-import org.globalforestwatch.features.FeatureId
-import org.globalforestwatch.util.Util.getAnyMapValue
-
-object AnnualUpdateAnalysis {
-  def apply(featureRDD: RDD[Feature[Geometry, FeatureId]],
-            featureType: String,
-            spark: SparkSession,
-            kwargs: Map[String, Any]): Unit = {
-
-    import spark.implicits._
-
-    val summaryRDD: RDD[(FeatureId, AnnualUpdateSummary)] =
-      AnnualUpdateRDD(featureRDD, AnnualUpdateGrid.blockTileGrid, kwargs)
-
-    val summaryDF =
-      AnnualUpdateDFFactory(featureType, summaryRDD, spark).getDataFrame
-
-    //    val maybeOutputPartitions:Option[Int] = getAnyMapValue(kwargs,"maybeOutputPartitions")
-    //    val outputPartitionCount =
-    //      maybeOutputPartitions.getOrElse(featureRDD.getNumPartitions)
-
-    summaryDF.repartition($"id", $"data_group")
-
-    val runOutputUrl: String = getAnyMapValue[String](kwargs, "outputUrl") +
-      "/annualupdate_" + DateTimeFormatter
-      .ofPattern("yyyyMMdd_HHmm")
-      .format(LocalDateTime.now)
-
-    AnnualUpdateExport.export(featureType, summaryDF, runOutputUrl, kwargs)
-  }
-}
