--- src/main/scala/org/globalforestwatch/summarystats/annualupdate/AnnualUpdateDataGroup.scala
@@ -1,43 +1,0 @@
-package org.globalforestwatch.summarystats.annualupdate
-
-case class AnnualUpdateDataGroup(lossYear: Integer,
-                                 threshold: Integer,
-                                 drivers: String,
-                                 globalLandCover: String,
-                                 primaryForest: Boolean,
-                                 idnPrimaryForest: Boolean,
-                                 erosion: String,
-                                 biodiversitySignificance: Boolean,
-                                 biodiversityIntactness: Boolean,
-                                 wdpa: String,
-                                 aze: Boolean,
-                                 plantations: String,
-                                 riverBasins: String,
-                                 ecozones: String,
-                                 urbanWatersheds: Boolean,
-                                 mangroves1996: Boolean,
-                                 mangroves2016: Boolean,
-                                 waterStress: String,
-                                 intactForestLandscapes: String,
-                                 endemicBirdAreas: Boolean,
-                                 tigerLandscapes: Boolean,
-                                 landmark: Boolean,
-                                 landRights: Boolean,
-                                 keyBiodiversityAreas: Boolean,
-                                 mining: Boolean,
-                                 rspo: String,
-                                 peatlands: Boolean,
-                                 oilPalm: Boolean,
-                                 idnForestMoratorium: Boolean,
-                                 idnLandCover: String,
-                                 mexProtectedAreas: Boolean,
-                                 mexPaymentForEcosystemServices: Boolean,
-                                 mexForestZoning: String,
-                                 perProductionForest: Boolean,
-                                 perProtectedAreas: Boolean,
-                                 perForestConcessions: String,
-                                 braBiomes: String,
-                                 woodFiber: Boolean,
-                                 resourceRights: Boolean,
-                                 logging: Boolean,
-                                 oilGas: Boolean)
