--- src/main/scala/org/globalforestwatch/summarystats/annualupdate/AnnualUpdateRDD.scala
@@ -1,38 +1,0 @@
-package org.globalforestwatch.summarystats.annualupdate
-
-import cats.implicits._
-import geotrellis.layer.{LayoutDefinition, SpatialKey}
-import geotrellis.raster.summary.polygonal._
-import geotrellis.raster.summary.GridVisitor
-import geotrellis.raster._
-import geotrellis.raster.rasterize.Rasterizer
-import geotrellis.vector._
-import org.globalforestwatch.summarystats.SummaryRDD
-
-object AnnualUpdateRDD extends SummaryRDD {
-
-  type SOURCES = AnnualUpdateGridSources
-  type SUMMARY = AnnualUpdateSummary
-  type TILE = AnnualUpdateTile
-
-  def getSources(windowKey: SpatialKey, windowLayout: LayoutDefinition, kwargs:Map[String, Any]): Either[Throwable, SOURCES] = {
-    Either.catchNonFatal {
-      AnnualUpdateGrid.getRasterSource(windowKey, windowLayout, kwargs)
-    }
-  }
-
-  def readWindow(rs: SOURCES, windowKey: SpatialKey, windowLayout: LayoutDefinition): Either[Throwable, Raster[TILE]] =
-    rs.readWindow(windowKey, windowLayout)
-
-  def runPolygonalSummary(raster: Raster[TILE],
-                          geometry: Geometry,
-                          options: Rasterizer.Options,
-                          kwargs: Map[String, Any]): PolygonalSummaryResult[SUMMARY] = {
-    raster.polygonalSummary(
-      geometry,
-      AnnualUpdateSummary.getGridVisitor(kwargs),
-      options = options
-    )
-  }
-
-}
