--- src/main/scala/org/globalforestwatch/summarystats/annualupdate/AnnualUpdateRow.scala
@@ -1,7 +1,0 @@
-package org.globalforestwatch.summarystats.annualupdate
-
-import org.globalforestwatch.features.GadmFeatureId
-
-case class AnnualUpdateRow(id: GadmFeatureId,
-                           dataGroup: AnnualUpdateDataGroup,
-                           data: AnnualUpdateData)
