+++ src/main/scala/org/globalforestwatch/summarystats/annualupdate_minimal/AnnualUpdateMinimalDF.scala
@@ -108,15 +108,15 @@ object AnnualUpdateMinimalDF {
         sum("umd_tree_cover_gain_2000-2012__ha") as "umd_tree_cover_gain_2000-2012__ha",
         sum("whrc_aboveground_biomass_stock_2000__Mg") as "whrc_aboveground_biomass_stock_2000__Mg",
         sum("whrc_aboveground_co2_stock_2000__Mg") as "whrc_aboveground_co2_stock_2000__Mg",
-        sum("umd_tree_cover_loss__ha") as "umd_tree_cover_loss_2001-2019__ha",
-        sum("whrc_aboveground_biomass_loss__Mg") as "whrc_aboveground_biomass_loss_2001-2019__Mg",
-        sum("gfw_gross_cumulative_aboveground_co2_removals__Mg") as "gfw_gross_cumulative_aboveground_co2_removals_2001-2019__Mg",
-        sum("gfw_gross_cumulative_belowground_co2_removals__Mg") as "gfw_gross_cumulative_belowground_co2_removals_2001-2019__Mg",
-        sum("gfw_gross_cumulative_aboveground_belowground_co2_removals__Mg") as "gfw_gross_cumulative_aboveground_belowground_co2_removals_2001-2019__Mg",
-        sum("gfw_net_flux_co2e__Mg") as "gfw_net_flux_co2e_2001-2019__Mg",
-        sum("gfw_gross_emissions_co2e_co2_only__Mg") as "gfw_gross_emissions_co2e_co2_only_2001-2019__Mg",
-        sum("gfw_gross_emissions_co2e_non_co2__Mg") as "gfw_gross_emissions_co2e_non_co2_2001-2019__Mg",
-        sum("gfw_gross_emissions_co2e_all_gases__Mg") as "gfw_gross_emissions_co2e_all_gases_2001-2019__Mg",
+        sum("umd_tree_cover_loss__ha") as "umd_tree_cover_loss_2001-2020__ha",
+        sum("whrc_aboveground_biomass_loss__Mg") as "whrc_aboveground_biomass_loss_2001-2020__Mg",
+        sum("gfw_gross_cumulative_aboveground_co2_removals__Mg") as "gfw_gross_cumulative_aboveground_co2_removals_2001-2020__Mg",
+        sum("gfw_gross_cumulative_belowground_co2_removals__Mg") as "gfw_gross_cumulative_belowground_co2_removals_2001-2020__Mg",
+        sum("gfw_gross_cumulative_aboveground_belowground_co2_removals__Mg") as "gfw_gross_cumulative_aboveground_belowground_co2_removals_2001-2020__Mg",
+        sum("gfw_net_flux_co2e__Mg") as "gfw_net_flux_co2e_2001-2020__Mg",
+        sum("gfw_gross_emissions_co2e_co2_only__Mg") as "gfw_gross_emissions_co2e_co2_only_2001-2020__Mg",
+        sum("gfw_gross_emissions_co2e_non_co2__Mg") as "gfw_gross_emissions_co2e_non_co2_2001-2020__Mg",
+        sum("gfw_gross_emissions_co2e_all_gases__Mg") as "gfw_gross_emissions_co2e_all_gases_2001-2020__Mg",
       )
   }
 
@@ -138,15 +138,15 @@ object AnnualUpdateMinimalDF {
         sum("umd_tree_cover_gain_2000-2012__ha") as "umd_tree_cover_gain_2000-2012__ha",
         sum("whrc_aboveground_biomass_stock_2000__Mg") as "whrc_aboveground_biomass_stock_2000__Mg",
         sum("whrc_aboveground_co2_stock_2000__Mg") as "whrc_aboveground_co2_stock_2000__Mg",
-        sum("umd_tree_cover_loss_2001-2019__ha") as "umd_tree_cover_loss_2001-2019__ha",
-        sum("whrc_aboveground_biomass_loss_2001-2019__Mg") as "whrc_aboveground_biomass_loss_2001-2019__Mg",
-        sum("gfw_gross_cumulative_aboveground_co2_removals_2001-2019__Mg") as "gfw_gross_cumulative_aboveground_co2_removals_2001-2019__Mg",
-        sum("gfw_gross_cumulative_belowground_co2_removals_2001-2019__Mg") as "gfw_gross_cumulative_belowground_co2_removals_2001-2019__Mg",
-        sum("gfw_gross_cumulative_aboveground_belowground_co2_removals_2001-2019__Mg") as "gfw_gross_cumulative_aboveground_belowground_co2_removals_2001-2019__Mg",
-        sum("gfw_net_flux_co2e_2001-2019__Mg") as "gfw_net_flux_co2e_2001-2019__Mg",
-        sum("gfw_gross_emissions_co2e_co2_only_2001-2019__Mg") as "gfw_gross_emissions_co2e_co2_only_2001-2019__Mg",
-        sum("gfw_gross_emissions_co2e_non_co2_2001-2019__Mg") as "gfw_gross_emissions_co2e_non_co2_2001-2019__Mg",
-        sum("gfw_gross_emissions_co2e_all_gases_2001-2019__Mg") as "gfw_gross_emissions_co2e_all_gases_2001-2019__Mg",
+        sum("umd_tree_cover_loss_2001-2020__ha") as "umd_tree_cover_loss_2001-2020__ha",
+        sum("whrc_aboveground_biomass_loss_2001-2020__Mg") as "whrc_aboveground_biomass_loss_2001-2020__Mg",
+        sum("gfw_gross_cumulative_aboveground_co2_removals_2001-2020__Mg") as "gfw_gross_cumulative_aboveground_co2_removals_2001-2020__Mg",
+        sum("gfw_gross_cumulative_belowground_co2_removals_2001-2020__Mg") as "gfw_gross_cumulative_belowground_co2_removals_2001-2020__Mg",
+        sum("gfw_gross_cumulative_aboveground_belowground_co2_removals_2001-2020__Mg") as "gfw_gross_cumulative_aboveground_belowground_co2_removals_2001-2020__Mg",
+        sum("gfw_net_flux_co2e_2001-2020__Mg") as "gfw_net_flux_co2e_2001-2020__Mg",
+        sum("gfw_gross_emissions_co2e_co2_only_2001-2020__Mg") as "gfw_gross_emissions_co2e_co2_only_2001-2020__Mg",
+        sum("gfw_gross_emissions_co2e_non_co2_2001-2020__Mg") as "gfw_gross_emissions_co2e_non_co2_2001-2020__Mg",
+        sum("gfw_gross_emissions_co2e_all_gases_2001-2020__Mg") as "gfw_gross_emissions_co2e_all_gases_2001-2020__Mg",
       )
   }
