+++ src/main/scala/org/globalforestwatch/summarystats/carbon_sensitivity/CarbonSensitivityDF.scala
@@ -27,7 +27,7 @@ object CarbonSensitivityDF {
       "bra_biome__name",
       "mapbox_river_basin__name",
       "is__regional_primary_forest",
-      "is__prodes_legal_amazon_umd_tree_cover_loss_2001-2019",
+      "is__prodes_legal_amazon_umd_tree_cover_loss_2001-2020",
       "is__prodes_legal_Amazon_extent_2000",
       "is__tropical_latitude_extent",
       "is__gfw_burn_loss_2001-19",
@@ -66,7 +66,7 @@ object CarbonSensitivityDF {
         $"dataGroup.braBiomes" as "bra_biome__name",
         $"dataGroup.riverBasins" as "mapbox_river_basin__name",
         $"dataGroup.primaryForest" as "is__regional_primary_forest",
-        $"dataGroup.isLossLegalAmazon" as "is__prodes_legal_amazon_umd_tree_cover_loss_2001-2019",
+        $"dataGroup.isLossLegalAmazon" as "is__prodes_legal_amazon_umd_tree_cover_loss_2001-2020",
         $"dataGroup.prodesLegalAmazonExtent2000" as "is__prodes_legal_Amazon_extent_2000",
         $"dataGroup.tropicLatitudeExtent" as "is__tropical_latitude_extent",
         $"dataGroup.isBurnLoss" as "is__gfw_burn_loss_2001-19",
@@ -99,17 +99,17 @@ object CarbonSensitivityDF {
           sum("umd_tree_cover_extent_2000__ha") as "umd_tree_cover_extent_2000__ha",
           sum("area__ha") as "area__ha",
           sum("whrc_aboveground_biomass_stock_2000__Mg") as "whrc_aboveground_biomass_stock_2000__Mg",
-          sum("gfw_gross_cumulative_aboveground_co2_removals__Mg") as "gfw_gross_cumulative_aboveground_co2_removals_2001-2019__Mg",
-          sum("gfw_gross_cumulative_belowground_co2_removals__Mg") as "gfw_gross_cumulative_belowground_co2_removals_2001-2019__Mg",
-          sum("gfw_gross_cumulative_aboveground_belowground_co2_removals__Mg") as "gfw_gross_cumulative_aboveground_belowground_co2_removals_2001-2019__Mg",
-          sum("gfw_net_flux_co2e__Mg") as "gfw_net_flux_co2e_2001-2019__Mg",
-          sum("umd_tree_cover_loss__ha") as "treecover_loss_2001-2019__ha",
-          sum("whrc_aboveground_biomass_loss__Mg") as "aboveground_biomass_loss_2001-2019__Mg",
-          sum("gfw_gross_emissions_co2e_co2_only__Mg") as "gfw_gross_emissions_co2e_co2_only_2001-2019__Mg",
-          sum("gfw_gross_emissions_co2e_non_co2__Mg") as "gfw_gross_emissions_co2e_non_co2_2001-2019__Mg",
-          sum("gfw_gross_emissions_co2e_all_gases__Mg") as "gfw_gross_emissions_co2e_all_gases_2001-2019__Mg",
+          sum("gfw_gross_cumulative_aboveground_co2_removals__Mg") as "gfw_gross_cumulative_aboveground_co2_removals_2001-2020__Mg",
+          sum("gfw_gross_cumulative_belowground_co2_removals__Mg") as "gfw_gross_cumulative_belowground_co2_removals_2001-2020__Mg",
+          sum("gfw_gross_cumulative_aboveground_belowground_co2_removals__Mg") as "gfw_gross_cumulative_aboveground_belowground_co2_removals_2001-2020__Mg",
+          sum("gfw_net_flux_co2e__Mg") as "gfw_net_flux_co2e_2001-2020__Mg",
+          sum("umd_tree_cover_loss__ha") as "treecover_loss_2001-2020__ha",
+          sum("whrc_aboveground_biomass_loss__Mg") as "aboveground_biomass_loss_2001-2020__Mg",
+          sum("gfw_gross_emissions_co2e_co2_only__Mg") as "gfw_gross_emissions_co2e_co2_only_2001-2020__Mg",
+          sum("gfw_gross_emissions_co2e_non_co2__Mg") as "gfw_gross_emissions_co2e_non_co2_2001-2020__Mg",
+          sum("gfw_gross_emissions_co2e_all_gases__Mg") as "gfw_gross_emissions_co2e_all_gases_2001-2020__Mg",
           sum("jpl_tropics_aboveground_biomass_2000__Mg") as "jpl_tropics_aboveground_biomass_2000__Mg",
-          sum("legal_amazon_umd_tree_cover_loss__ha") as "legal_amazon_umd_tree_cover_loss_2001-2019__ha",
+          sum("legal_amazon_umd_tree_cover_loss__ha") as "legal_amazon_umd_tree_cover_loss_2001-2020__ha",
           sum("gfw_flux_model_extent__ha") as "gfw_flux_model_extent__ha"
 
         )
@@ -122,17 +122,17 @@ object CarbonSensitivityDF {
           sum("umd_tree_cover_extent_2000__ha") as "umd_tree_cover_extent_2000__ha",
           sum("area__ha") as "area__ha",
           sum("whrc_aboveground_biomass_stock_2000__Mg") as "whrc_aboveground_biomass_stock_2000__Mg",
-          sum("gfw_gross_cumulative_aboveground_co2_removals_2001-2019__Mg") as "gfw_gross_cumulative_aboveground_co2_removals_2001-2019__Mg",
-          sum("gfw_gross_cumulative_belowground_co2_removals_2001-2019__Mg") as "gfw_gross_cumulative_belowground_co2_removals_2001-2019__Mg",
-          sum("gfw_gross_cumulative_aboveground_belowground_co2_removals_2001-2019__Mg") as "gfw_gross_cumulative_aboveground_belowground_co2_removals_2001-2019__Mg",
-          sum("gfw_net_flux_co2e_2001-2019__Mg") as "gfw_net_flux_co2e_2001-2019__Mg",
-          sum("treecover_loss_2001-2019__ha") as "treecover_loss_2001-2019__ha",
-          sum("aboveground_biomass_loss_2001-2019__Mg") as "aboveground_biomass_loss_2001-2019__Mg",
-          sum("gfw_gross_emissions_co2e_co2_only_2001-2019__Mg") as "gfw_gross_emissions_co2e_co2_only_2001-2019__Mg",
-          sum("gfw_gross_emissions_co2e_non_co2_2001-2019__Mg") as "gfw_gross_emissions_co2e_non_co2_2001-2019__Mg",
-          sum("gfw_gross_emissions_co2e_all_gases_2001-2019__Mg") as "gfw_gross_emissions_co2e_all_gases_2001-2019__Mg",
+          sum("gfw_gross_cumulative_aboveground_co2_removals_2001-2020__Mg") as "gfw_gross_cumulative_aboveground_co2_removals_2001-2020__Mg",
+          sum("gfw_gross_cumulative_belowground_co2_removals_2001-2020__Mg") as "gfw_gross_cumulative_belowground_co2_removals_2001-2020__Mg",
+          sum("gfw_gross_cumulative_aboveground_belowground_co2_removals_2001-2020__Mg") as "gfw_gross_cumulative_aboveground_belowground_co2_removals_2001-2020__Mg",
+          sum("gfw_net_flux_co2e_2001-2020__Mg") as "gfw_net_flux_co2e_2001-2020__Mg",
+          sum("treecover_loss_2001-2020__ha") as "treecover_loss_2001-2020__ha",
+          sum("aboveground_biomass_loss_2001-2020__Mg") as "aboveground_biomass_loss_2001-2020__Mg",
+          sum("gfw_gross_emissions_co2e_co2_only_2001-2020__Mg") as "gfw_gross_emissions_co2e_co2_only_2001-2020__Mg",
+          sum("gfw_gross_emissions_co2e_non_co2_2001-2020__Mg") as "gfw_gross_emissions_co2e_non_co2_2001-2020__Mg",
+          sum("gfw_gross_emissions_co2e_all_gases_2001-2020__Mg") as "gfw_gross_emissions_co2e_all_gases_2001-2020__Mg",
           sum("jpl_tropics_aboveground_biomass_2000__Mg") as "jpl_tropics_aboveground_biomass_2000__Mg",
-          sum("legal_amazon_umd_tree_cover_loss_2001-2019__ha") as "legal_amazon_umd_tree_cover_loss_2001-2019__ha",
+          sum("legal_amazon_umd_tree_cover_loss_2001-2020__ha") as "legal_amazon_umd_tree_cover_loss_2001-2020__ha",
           sum("gfw_flux_model_extent__ha") as "gfw_flux_model_extent__ha"
 
         )
@@ -181,7 +181,7 @@ object CarbonSensitivityDF {
           max(length($"bra_biome__name")).cast("boolean") as "bra_biome__name",
           max(length($"mapbox_river_basin__name")).cast("boolean") as "mapbox_river_basin__name",
           max($"is__regional_primary_forest") as "is__regional_primary_forest",
-          max($"is__prodes_legal_amazon_umd_tree_cover_loss_2001-2019") as "is__prodes_legal_amazon_umd_tree_cover_loss_2001-2019",
+          max($"is__prodes_legal_amazon_umd_tree_cover_loss_2001-2020") as "is__prodes_legal_amazon_umd_tree_cover_loss_2001-2020",
           max($"is__prodes_legal_Amazon_extent_2000") as "is__prodes_legal_Amazon_extent_2000",
           max($"is__tropical_latitude_extent") as "is__tropical_latitude_extent",
           max($"is__gfw_burn_loss_2001-19") as "is__gfw_burn_loss_2001-19",
@@ -215,7 +215,7 @@ object CarbonSensitivityDF {
           max($"bra_biome__name") as "bra_biome__name",
           max($"mapbox_river_basin__name") as "mapbox_river_basin__name",
           max($"is__regional_primary_forest") as "is__regional_primary_forest",
-          max($"is__prodes_legal_amazon_umd_tree_cover_loss_2001-2019") as "is__prodes_legal_amazon_umd_tree_cover_loss_2001-2019",
+          max($"is__prodes_legal_amazon_umd_tree_cover_loss_2001-2020") as "is__prodes_legal_amazon_umd_tree_cover_loss_2001-2020",
           max($"is__prodes_legal_Amazon_extent_2000") as "is__prodes_legal_Amazon_extent_2000",
           max($"is__tropical_latitude_extent") as "is__tropical_latitude_extent",
           max($"is__gfw_burn_loss_2001-19") as "is__gfw_burn_loss_2001-19",
