--- src/main/scala/org/globalforestwatch/summarystats/carbonflux_minimal/CarbonFluxMinimalDataGroup.scala
@@ -1,8 +1,0 @@
-package org.globalforestwatch.summarystats.carbonflux_minimal
-
-case class CarbonFluxMinimalDataGroup(
-                              threshold: Integer,
-                              tcdYear: Int,
-                              isPrimaryForest: Boolean,
-                              isPlantations: Boolean
-                        )
