--- src/main/scala/org/globalforestwatch/summarystats/carbonflux_minimal/CarbonFluxMinimalRDD.scala
@@ -1,39 +1,0 @@
-package org.globalforestwatch.summarystats.carbonflux_minimal
-
-import cats.implicits._
-import geotrellis.layer.{LayoutDefinition, SpatialKey}
-import geotrellis.raster.summary.polygonal._
-import geotrellis.raster.summary.GridVisitor
-import geotrellis.raster._
-import geotrellis.raster.rasterize.Rasterizer
-import geotrellis.vector._
-import org.globalforestwatch.summarystats.SummaryRDD
-
-object CarbonFluxMinimalRDD extends SummaryRDD {
-
-  type SOURCES = CarbonFluxMinimalGridSources
-  type SUMMARY = CarbonFluxMinimalSummary
-  type TILE = CarbonFluxMinimalTile
-
-  def getSources(windowKey: SpatialKey, windowLayout: LayoutDefinition, kwargs: Map[String, Any]): Either[Throwable, SOURCES] = {
-    Either.catchNonFatal {
-      CarbonFluxMinimalGrid.getRasterSource(windowKey, windowLayout, kwargs)
-    }
-  }
-
-  def readWindow(rs: SOURCES,
-                 windowKey: SpatialKey, windowLayout: LayoutDefinition): Either[Throwable, Raster[TILE]] =
-    rs.readWindow(windowKey, windowLayout)
-
-  def runPolygonalSummary(raster: Raster[TILE],
-                          geometry: Geometry,
-                          options: Rasterizer.Options,
-                          kwargs: Map[String, Any]): PolygonalSummaryResult[SUMMARY] = {
-    raster.polygonalSummary(
-      geometry,
-      CarbonFluxMinimalSummary.getGridVisitor(kwargs),
-      options = options
-    )
-  }
-
-}
