--- src/main/scala/org/globalforestwatch/summarystats/carbonflux_minimal/CarbonFluxMinimalRowGadm.scala
@@ -1,7 +1,0 @@
-package org.globalforestwatch.summarystats.carbonflux_minimal
-
-import org.globalforestwatch.features.GadmFeatureId
-
-case class CarbonFluxMinimalRowGadm(id: GadmFeatureId,
-                                    dataGroup: CarbonFluxMinimalDataGroup,
-                                    data: CarbonFluxMinimalData)
