--- src/main/scala/org/globalforestwatch/summarystats/carbonflux_minimal/CarbonFluxMinimalRowSimple.scala
@@ -1,7 +1,0 @@
-package org.globalforestwatch.summarystats.carbonflux_minimal
-
-import org.globalforestwatch.features.SimpleFeatureId
-
-case class CarbonFluxMinimalRowSimple(id: SimpleFeatureId,
-                                      dataGroup: CarbonFluxMinimalDataGroup,
-                                      data: CarbonFluxMinimalData)
