--- src/main/scala/org/globalforestwatch/summarystats/carbonflux_minimal/CarbonFluxMinimalRowWdpa.scala
@@ -1,7 +1,0 @@
-package org.globalforestwatch.summarystats.carbonflux_minimal
-
-import org.globalforestwatch.features.WdpaFeatureId
-
-case class CarbonFluxMinimalRowWdpa(id: WdpaFeatureId,
-                                    dataGroup: CarbonFluxMinimalDataGroup,
-                                    data: CarbonFluxMinimalData)
