+++ src/main/scala/org/globalforestwatch/summarystats/treecoverloss/TreeLossSummary.scala
@@ -45,6 +45,13 @@ object TreeLossSummary {
         val tcd2010: Integer = raster.tile.tcd2010.getData(col, row)
         val biomass: Double = raster.tile.biomass.getData(col, row)
 
+        val grossCumulAbovegroundRemovalsCo2: Float = raster.tile.grossCumulAbovegroundRemovalsCo2.getData(col, row)
+        val grossCumulBelowgroundRemovalsCo2: Float = raster.tile.grossCumulBelowgroundRemovalsCo2.getData(col, row)
+        val grossEmissionsCo2eNonCo2: Float = raster.tile.grossEmissionsCo2eNonCo2.getData(col, row)
+        val grossEmissionsCo2eCo2Only: Float =  raster.tile.grossEmissionsCo2eCo2Only.getData(col, row)
+        val netFluxCo2: Float = raster.tile.netFluxCo2.getData(col, row)
+        val fluxModelExtent: Boolean = raster.tile.fluxModelExtent.getData(col, row)
+
         val contextualLayers: List[String] =
           getAnyMapValue[NonEmptyList[String]](kwargs, "contextualLayers").toList
 
@@ -72,6 +79,19 @@ object TreeLossSummary {
         val biomassPixel = biomass * areaHa
         val co2Pixel = biomassPixel * co2Factor
 
+        val grossCumulAbovegroundRemovalsCo2Pixel = grossCumulAbovegroundRemovalsCo2 * areaHa
+        val grossCumulBelowgroundRemovalsCo2Pixel = grossCumulBelowgroundRemovalsCo2 * areaHa
+        val grossCumulAboveBelowgroundRemovalsCo2Pixel = grossCumulAbovegroundRemovalsCo2Pixel + grossCumulBelowgroundRemovalsCo2Pixel
+
+        val grossEmissionsCo2eNonCo2Pixel = grossEmissionsCo2eNonCo2 * areaHa
+        val grossEmissionsCo2eCo2OnlyPixel = grossEmissionsCo2eCo2Only * areaHa
+        val grossEmissionsCo2eAllGasesPixel = grossEmissionsCo2eNonCo2Pixel + grossEmissionsCo2eCo2OnlyPixel
+
+        val netFluxCo2Pixel = netFluxCo2 * areaHa
+
+        val fluxModelExtentAreaInt: Integer = if (fluxModelExtent) 1 else 0
+        val fluxModelExtentAreaPixel: Double = fluxModelExtentAreaInt * areaHa
+
         val thresholds: List[Int] =
            getAnyMapValue[NonEmptyList[Int]](kwargs, "thresholdFilter").toList
 
@@ -94,7 +114,7 @@ object TreeLossSummary {
               stats.getOrElse(
                 key = pKey,
                 default =
-                  TreeLossData(TreeLossYearDataMap.empty, 0, 0, 0, 0, 0, 0, 0)
+                  TreeLossData(TreeLossYearDataMap.empty, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0)
               )
 
             summary.totalArea += areaHa
@@ -107,6 +127,9 @@ object TreeLossSummary {
                 summary.lossYear(loss).treecoverLoss += areaHa
                 summary.lossYear(loss).biomassLoss += biomassPixel
                 summary.lossYear(loss).carbonEmissions += co2Pixel
+                summary.lossYear(loss).grossEmissionsCo2eCo2Only += grossEmissionsCo2eCo2OnlyPixel
+                summary.lossYear(loss).grossEmissionsCo2eNonCo2 += grossEmissionsCo2eNonCo2Pixel
+                summary.lossYear(loss).grossEmissionsCo2eAllGases += grossEmissionsCo2eAllGasesPixel
               }
 
               // TODO: use extent2010 to calculate avg biomass incase year is selected
@@ -117,6 +140,18 @@ object TreeLossSummary {
               }
               summary.totalBiomass += biomassPixel
               summary.totalCo2 += co2Pixel
+
+              summary.totalGrossCumulAbovegroundRemovalsCo2 += grossCumulAbovegroundRemovalsCo2Pixel
+              summary.totalGrossCumulBelowgroundRemovalsCo2 += grossCumulBelowgroundRemovalsCo2Pixel
+              summary.totalGrossCumulAboveBelowgroundRemovalsCo2 += grossCumulAboveBelowgroundRemovalsCo2Pixel
+
+              summary.totalGrossEmissionsCo2eCo2Only += grossEmissionsCo2eCo2OnlyPixel
+              summary.totalGrossEmissionsCo2eNonCo2 += grossEmissionsCo2eNonCo2Pixel
+              summary.totalGrossEmissionsCo2eAllGases += grossEmissionsCo2eAllGasesPixel
+
+              summary.totalNetFluxCo2 += netFluxCo2Pixel
+
+              summary.totalFluxModelExtentArea += fluxModelExtentAreaPixel
             }
 
             tcdYear match {
