+++ src/main/scala/org/globalforestwatch/summarystats/treecoverloss/TreeLossTile.scala
@@ -14,7 +14,13 @@ case class TreeLossTile(loss: TreeCoverLoss#ITile,
                         tcd2010: TreeCoverDensity2010#ITile,
                         biomass: BiomassPerHectar#OptionalDTile,
                         primaryForest: PrimaryForest#OptionalITile,
-                        plantationsBool: PlantationsBool#OptionalITile)
+                        plantationsBool: PlantationsBool#OptionalITile,
+                        grossCumulAbovegroundRemovalsCo2: GrossCumulAbovegroundRemovalsCo2#OptionalFTile,
+                        grossCumulBelowgroundRemovalsCo2: GrossCumulBelowgroundRemovalsCo2#OptionalFTile,
+                        netFluxCo2: NetFluxCo2e#OptionalFTile,
+                        grossEmissionsCo2eNonCo2: GrossEmissionsNonCo2Co2e#OptionalFTile,
+                        grossEmissionsCo2eCo2Only: GrossEmissionsCo2OnlyCo2e#OptionalFTile,
+                        fluxModelExtent: FluxModelExtent#OptionalITile)
 
   extends CellGrid[Int] {
